change_text e change_text_failed
change_text script is used to make writing appear between levels depending on the type of person, if a person if a person passes the level. change_text script is used to make writing appear between levels depending on the type of person, if a person if a person passes failed the level.